requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2
pydantic-settings==2.0.3
python-multipart==0.0.6

# AI/ML Libraries
torch==2.1.0
transformers==4.35.0
sentence-transformers==2.2.2
openai==1.3.0
anthropic==0.7.0
google-generativeai==0.3.0
tiktoken==0.5.1

# Computer Vision
opencv-python==4.8.1.78
pillow==10.0.1
scikit-image==0.22.0

# Data Processing
numpy==1.24.3
pandas==2.0.3
scipy==1.11.4
scikit-learn==1.3.2

# Database & Storage
redis==5.0.1
psycopg2-binary==2.9.7
pymongo==4.5.0
chromadb==0.4.15
sqlalchemy==2.0.23
alembic==1.12.1

# Message Queue & Task Management
kafka-python==2.0.2
celery==5.3.4
kombu==5.3.3

# Vector Operations
faiss-cpu==1.7.4
annoy==1.17.3

# HTTP & Networking
httpx==0.25.1
aiohttp==3.9.0
websockets==12.0
python-socketio==5.10.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
click==8.1.7
rich==13.6.0
tqdm==4.66.1
python-json-logger==2.0.7

# Monitoring & Logging
prometheus-client==0.19.0
structlog==23.2.0

# Caching
cachetools==5.3.2
diskcache==5.6.3

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==41.0.7

# Date & Time
python-dateutil==2.8.2
pytz==2023.3

# Configuration Management
hydra-core==1.3.2
omegaconf==2.3.0

# API Documentation
fastapi-users==12.1.2

# Async Support
asyncio==3.4.3
aiofiles==23.2.1

# Graph Operations (for reasoning)
networkx==3.2.1

# NLP Utilities
spacy==3.7.2
nltk==3.8.1
textblob==0.17.1

# Math & Statistics
sympy==1.12
statsmodels==0.14.0